fix placeholders for more than one locale (#35)
The final_attrs dict is reused across all locales, so the 2nd and all
later locales already have a placeholder set. This causes the setdefault
to do nothing, as intended. This changes the structure to create a fresh
dict for every locale, based on the final_attrs generated by django.